Refactor lr package
