--- parser/lr/calculator_test.go
@@ -1,178 +1,0 @@
-package lr
-
-import (
-	"testing"
-
-	"github.com/stretchr/testify/assert"
-
-	"github.com/moorara/algo/grammar"
-)
-
-func TestCalculator0_G(t *testing.T) {
-	tests := []struct {
-		name string
-		c    *calculator0
-	}{
-		{
-			name: "OK",
-			c: &calculator0{
-				augG: augment(grammars[2]),
-			},
-		},
-	}
-
-	for _, tc := range tests {
-		t.Run(tc.name, func(t *testing.T) {
-			assert.NoError(t, tc.c.augG.Verify())
-			G := tc.c.G()
-
-			assert.True(t, G.Equal(tc.c.augG))
-		})
-	}
-}
-
-func TestCalculator0_Initial(t *testing.T) {
-	tests := []struct {
-		name            string
-		c               *calculator0
-		expectedInitial Item
-	}{
-		{
-			name: "OK",
-			c: &calculator0{
-				augG: augment(grammars[2]),
-			},
-			expectedInitial: &Item0{
-				Production: prods[2][0],
-				Start:      starts[2],
-				Dot:        0,
-			},
-		},
-	}
-
-	for _, tc := range tests {
-		t.Run(tc.name, func(t *testing.T) {
-			assert.NoError(t, tc.c.augG.Verify())
-			initial := tc.c.Initial()
-
-			assert.True(t, initial.Equal(tc.expectedInitial))
-		})
-	}
-}
-
-func TestCalculator0_CLOSURE(t *testing.T) {
-	s := getTestLR0ItemSets()
-
-	tests := []struct {
-		name            string
-		c               *calculator0
-		I               ItemSet
-		expectedCLOSURE ItemSet
-	}{
-		{
-			name: "OK",
-			c: &calculator0{
-				augG: augment(grammars[2]),
-			},
-			I: NewItemSet(
-				&Item0{Production: prods[2][0], Start: starts[2], Dot: 0}, // E′ → •E
-			),
-			expectedCLOSURE: s[0],
-		},
-	}
-
-	for _, tc := range tests {
-		t.Run(tc.name, func(t *testing.T) {
-			assert.NoError(t, tc.c.augG.Verify())
-			J := tc.c.CLOSURE(tc.I)
-
-			assert.True(t, J.Equal(tc.expectedCLOSURE))
-		})
-	}
-}
-
-func TestCalculator1_G(t *testing.T) {
-	tests := []struct {
-		name string
-		c    *calculator1
-	}{
-		{
-			name: "OK",
-			c: &calculator1{
-				augG: augment(grammars[0]),
-			},
-		},
-	}
-
-	for _, tc := range tests {
-		t.Run(tc.name, func(t *testing.T) {
-			assert.NoError(t, tc.c.augG.Verify())
-			G := tc.c.G()
-
-			assert.True(t, G.Equal(tc.c.augG))
-		})
-	}
-}
-
-func TestCalculator1_Initial(t *testing.T) {
-	tests := []struct {
-		name            string
-		c               *calculator1
-		expectedInitial Item
-	}{
-		{
-			name: "OK",
-			c: &calculator1{
-				augG: augment(grammars[0]),
-			},
-			expectedInitial: &Item1{
-				Production: prods[0][0],
-				Start:      starts[0],
-				Dot:        0,
-				Lookahead:  grammar.Endmarker,
-			},
-		},
-	}
-
-	for _, tc := range tests {
-		t.Run(tc.name, func(t *testing.T) {
-			assert.NoError(t, tc.c.augG.Verify())
-			initial := tc.c.Initial()
-
-			assert.True(t, initial.Equal(tc.expectedInitial))
-		})
-	}
-}
-
-func TestCalculator1_CLOSURE(t *testing.T) {
-	s := getTestLR1ItemSets()
-	g := augment(grammars[0])
-
-	tests := []struct {
-		name            string
-		c               *calculator1
-		I               ItemSet
-		expectedCLOSURE ItemSet
-	}{
-		{
-			name: "OK",
-			c: &calculator1{
-				augG:  g,
-				FIRST: g.ComputeFIRST(),
-			},
-			I: NewItemSet(
-				&Item1{Production: prods[0][0], Start: starts[0], Dot: 0, Lookahead: grammar.Endmarker}, // S′ → •S, $
-			),
-			expectedCLOSURE: s[0],
-		},
-	}
-
-	for _, tc := range tests {
-		t.Run(tc.name, func(t *testing.T) {
-			assert.NoError(t, tc.c.augG.Verify())
-			J := tc.c.CLOSURE(tc.I)
-
-			assert.True(t, J.Equal(tc.expectedCLOSURE))
-		})
-	}
-}
