+++ parser/lr/item.go
@@ -3,6 +3,8 @@ package lr
 import (
 	"bytes"
 	"fmt"
+	"strings"
+	"unicode/utf8"
 
 	"github.com/moorara/algo/generic"
 	"github.com/moorara/algo/grammar"
@@ -22,63 +24,9 @@ var (
 	EqItemSet = func(lhs, rhs ItemSet) bool {
 		return lhs.Equal(rhs)
 	}
+
+	CmpItemSet = cmpItemSet
 )
-
-// ItemSet represents a collection of item sets for a context-free grammar.
-type ItemSetCollection set.Set[ItemSet]
-
-// NewItemSetCollection creates a new collection of item sets.
-func NewItemSetCollection(sets ...ItemSet) ItemSetCollection {
-	return set.NewWithFormat(
-		EqItemSet,
-		func(sets []ItemSet) string {
-			cs := &itemSetCollectionStringer{sets: sets}
-			return cs.String()
-		},
-		sets...,
-	)
-}
-
-// ItemSet represents a set of items for a context-free grammar.
-type ItemSet set.Set[Item]
-
-// NewItemSet creates a new set of items.
-func NewItemSet(items ...Item) ItemSet {
-	return set.NewWithFormat(
-		EqItem,
-		func(items []Item) string {
-			ss := &itemSetStringer{items: items}
-			return ss.String()
-		},
-		items...,
-	)
-}
-
-// cmpItemSet compares two collections of item sets to establish a consistent and deterministic order.
-// The comparison process is as follows:
-//
-//  1. Each item set is sorted using cmpItem first.
-//  2. The items in the two lists are compared one by one in order using cmpItem.
-//  3. If one list is longer and all compared items are equal, the longer list comes first.
-//
-// This function is useful for sorting collections of item sets,
-// ensuring stability and predictability in their ordering.
-func cmpItemSet(lhs, rhs ItemSet) int {
-	ls := generic.Collect1(lhs.All())
-	sort.Quick(ls, CmpItem)
-
-	rs := generic.Collect1(rhs.All())
-	sort.Quick(rs, CmpItem)
-
-	for i := 0; i < len(ls) && i < len(rs); i++ {
-		if cmp := CmpItem(ls[i], rs[i]); cmp != 0 {
-			return cmp
-		}
-	}
-
-	//  If all compared items are equal, the longer list comes first.
-	return len(rs) - len(ls)
-}
 
 // Item represents an LR item for a context-free grammar.
 //
@@ -456,4 +404,202 @@ func (i *Item1) Item0() *Item0 {
 		Start:      i.Start,
 		Dot:        i.Dot,
 	}
+}
+
+// ItemSet represents a set of items for a context-free grammar.
+type ItemSet set.Set[Item]
+
+// NewItemSet creates a new set of items.
+func NewItemSet(items ...Item) ItemSet {
+	return set.NewWithFormat(
+		EqItem,
+		func(items []Item) string {
+			ss := &itemSetStringer{items: items}
+			return ss.String()
+		},
+		items...,
+	)
+}
+
+// cmpItemSet compares two collections of item sets to establish a consistent and deterministic order.
+// The comparison process is as follows:
+//
+//  1. Each item set is sorted using cmpItem first.
+//  2. The items in the two lists are compared one by one in order using cmpItem.
+//  3. If one list is longer and all compared items are equal, the longer list comes first.
+//
+// This function is useful for sorting collections of item sets,
+// ensuring stability and predictability in their ordering.
+func cmpItemSet(lhs, rhs ItemSet) int {
+	ls := generic.Collect1(lhs.All())
+	sort.Quick(ls, CmpItem)
+
+	rs := generic.Collect1(rhs.All())
+	sort.Quick(rs, CmpItem)
+
+	for i := 0; i < len(ls) && i < len(rs); i++ {
+		if cmp := CmpItem(ls[i], rs[i]); cmp != 0 {
+			return cmp
+		}
+	}
+
+	//  If all compared items are equal, the longer list comes first.
+	return len(rs) - len(ls)
+}
+
+// itemSetStringer builds a string representation of an item set.
+// It provides a human-readable visualization of a set of items.
+type itemSetStringer struct {
+	state *State
+	items []Item
+
+	b     bytes.Buffer
+	width int
+}
+
+func (s *itemSetStringer) String() string {
+	s.b.Reset()
+
+	// Calculate the width of item set.
+	s.calculateWidth()
+
+	sort.Quick(s.items, CmpItem)
+
+	s.printTopLine() // ┌────────[s]─────────┐
+
+	sepPrinted := false
+	for _, item := range s.items {
+		if !item.IsKernel() && !sepPrinted {
+			s.printMiddleLine()
+			sepPrinted = true
+		}
+		s.printItem(item)
+	}
+
+	s.printBottomLine() // └────────────────────┘
+
+	return s.b.String()
+}
+
+func (s *itemSetStringer) calculateWidth() {
+	// Already set or calculated.
+	if s.width != 0 {
+		return
+	}
+
+	// Minimum width
+	s.width = 10
+
+	for _, item := range s.items {
+		if l := utf8.RuneCountInString(item.String()); l > s.width {
+			s.width = l
+		}
+	}
+
+	// Padding
+	s.width += 2
+}
+
+func (s *itemSetStringer) printTopLine() {
+	var state string
+	if s.state != nil {
+		state = fmt.Sprintf("[%d]", *s.state)
+	}
+
+	pad := s.width - utf8.RuneCountInString(state)
+	lpad := pad / 2
+	rpad := pad - lpad
+
+	s.b.WriteRune('┌')
+	s.b.WriteString(strings.Repeat("─", lpad))
+	s.b.WriteString(state)
+	s.b.WriteString(strings.Repeat("─", rpad))
+	s.b.WriteRune('┐')
+	s.b.WriteRune('\n')
+}
+
+func (s *itemSetStringer) printItem(i Item) {
+	item := i.String()
+
+	rpad := s.width - utf8.RuneCountInString(item) - 2
+
+	s.b.WriteRune('│')
+	s.b.WriteRune(' ')
+	s.b.WriteString(item)
+	s.b.WriteString(strings.Repeat(" ", rpad))
+	s.b.WriteRune(' ')
+	s.b.WriteRune('│')
+	s.b.WriteRune('\n')
+}
+
+func (s *itemSetStringer) printMiddleLine() {
+	s.b.WriteRune('├')
+	s.b.WriteString(strings.Repeat("╌", s.width))
+	s.b.WriteRune('┤')
+	s.b.WriteRune('\n')
+}
+
+func (s *itemSetStringer) printBottomLine() {
+	s.b.WriteRune('└')
+	s.b.WriteString(strings.Repeat("─", s.width))
+	s.b.WriteRune('┘')
+	s.b.WriteRune('\n')
+}
+
+// ItemSet represents a collection of item sets for a context-free grammar.
+type ItemSetCollection set.Set[ItemSet]
+
+// NewItemSetCollection creates a new collection of item sets.
+func NewItemSetCollection(sets ...ItemSet) ItemSetCollection {
+	return set.NewWithFormat(
+		EqItemSet,
+		func(sets []ItemSet) string {
+			cs := &itemSetCollectionStringer{sets: sets}
+			return cs.String()
+		},
+		sets...,
+	)
+}
+
+// itemSetCollectionStringer builds a string representation of a collection of item sets.
+// It provides a human-readable visualization of a collection of item sets.
+type itemSetCollectionStringer struct {
+	sets []ItemSet
+}
+
+func (c *itemSetCollectionStringer) String() string {
+	var b bytes.Buffer
+
+	// Calculate the maximum length for all item sets.
+	maxWidth := c.calculateMaxWidth()
+
+	for i, set := range c.sets {
+		is := &itemSetStringer{
+			state: (*State)(&i),
+			items: generic.Collect1(set.All()),
+			width: maxWidth,
+		}
+
+		b.WriteString(is.String())
+	}
+
+	return b.String()
+}
+
+func (c *itemSetCollectionStringer) calculateMaxWidth() int {
+	// Minimum width
+	maxWidth := 10
+
+	for _, set := range c.sets {
+		for item := range set.All() {
+			if l := utf8.RuneCountInString(item.String()); l > maxWidth {
+				maxWidth = l
+			}
+		}
+	}
+
+	// Padding
+	maxWidth += 2
+
+	return maxWidth
 }
